requirements.txt
flask==2.3.3
apscheduler==3.10.4
pyyaml==6.0.1
imapclient==2.3.1
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0